+++ src/rt/boxed_region.cpp
@@ -1,6 +1,8 @@
-#include <assert.h>
+
+
 #include "boxed_region.h"
-#include "rust_internal.h"
+#include "rust_globals.h"
+#include "rust_task.h"
 
 // #define DUMP_BOXED_REGION
