+++ src/rt/rust.cpp
@@ -1,7 +1,8 @@
-#include "rust_internal.h"
+
+#include "rust_globals.h"
+#include "rust_kernel.h"
 #include "rust_util.h"
 #include "rust_scheduler.h"
-#include <cstdio>
 
 struct
 command_line_args : public kernel_owned<command_line_args>
