+++ src/rt/rust_kernel.cpp
@@ -1,4 +1,7 @@
-#include "rust_internal.h"
+
+#include "rust_kernel.h"
+#include "rust_globals.h"
+#include "rust_port.h"
 #include "rust_util.h"
 #include "rust_scheduler.h"
